requirements.txt
fastapi==0.124.0
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg[binary]==3.2.9
pgvector==0.2.4
openai==2.9.0
python-dotenv==1.0.0
pydantic==2.12.5
python-multipart==0.0.6
alembic==1.17.2
asyncpg==0.31.0
